Refactor SEO detail management by restructuring resources
- Moved SeoDetailResource to a new namespace: Firefly\FilamentBlog\Resources\SeoDetails.
- Introduced new schemas and pages for managing SEO details, including CreateSeoDetail, EditSeoDetail, and ListSeoDetails.
- Removed the deprecated SeoDetailResource class and its associated methods to enhance clarity and maintainability.
- Updated Blog.php and other related files to reflect the new SeoDetailResource path.
