+++ src/Blog.php
@@ -5,7 +5,7 @@ namespace Firefly\FilamentBlog;
 use Firefly\FilamentBlog\Resources\Categories\CategoryResource;
 use Firefly\FilamentBlog\Resources\Posts\PostResource;
 use Firefly\FilamentBlog\Resources\Tags\TagResource;
-use Firefly\FilamentBlog\Resources\SeoDetailResource;
+use Firefly\FilamentBlog\Resources\SeoDetails\SeoDetailResource;
 use Firefly\FilamentBlog\Resources\NewsletterResource;
 use Firefly\FilamentBlog\Resources\CommentResource;
 use Firefly\FilamentBlog\Resources\ShareSnippetResource;
