+++ src/Resources/SeoDetails/Pages/CreateSeoDetail.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Firefly\FilamentBlog\Resources\SeoDetailResource\Pages;
+namespace Firefly\FilamentBlog\Resources\SeoDetails\Pages;
 
 use Filament\Resources\Pages\CreateRecord;
-use Firefly\FilamentBlog\Resources\SeoDetailResource;
+use Firefly\FilamentBlog\Resources\SeoDetails\SeoDetailResource;
 
 class CreateSeoDetail extends CreateRecord
 {
