+++ src/Resources/SeoDetails/Pages/EditSeoDetail.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace Firefly\FilamentBlog\Resources\SeoDetailResource\Pages;
+namespace Firefly\FilamentBlog\Resources\SeoDetails\Pages;
 
 use Filament\Actions\DeleteAction;
-use Filament\Actions;
 use Filament\Resources\Pages\EditRecord;
-use Firefly\FilamentBlog\Resources\SeoDetailResource;
+use Firefly\FilamentBlog\Resources\SeoDetails\SeoDetailResource;
 
 class EditSeoDetail extends EditRecord
 {
